Append Principal Components to Data.Frame

Additional details about PCA forthcoming...

Usage

append_pc_to_df(data, var = NULL, std = FALSE)

Arguments

data

a data.frame or survey.design object

var

a character vector of variables names in data (default uses all variables in data)

std

a logical (defaults FALSE) of whether to standardized all resulting PCA after conducting PCA.

Value

the data.frame or survey.design object with PCs added to data (or under variables)

Details

Adds the principal components of a set of variables to the data.frame.

Examples

{
  # TO-DO
}
#> NULL