Create Keep Variable Vector

keep_variable(.), a wrapper around the function 'check_variance(.)' to create a logical vector of length(x) of the results of the check_variance function. The result can then be used to to subset the vector 'x', e.g., 'x[keep_variable(x, data)]', to only include predictors 'x' that have any variance. The inverse selection, 'x[!(keep_variable(x, data))]', can be used to identify which variables were excluded.

Usage

keep_variable(x, data)

Arguments

x

(character vector) of a vector of variable names in data

data

(data.frame)

Value

a logical vector of length equal to the length of 'x'

Details

Creates a logical vector to keep or delete variables for use in a regression model.

Examples

{
  df <- data.frame(A = c(1, 1, 0, 0), B = c("B", "B", "B", "B"))
  keep_variable(c("A", "B"), df)
}
#> [1]  TRUE FALSE