Create Predictor Matrix (modified)

Function copied and modified from mice pacakge (https://github.com/amices/mice/blob/master/R/quickpred.R)

Usage

quickpred2(
  data,
  mincor = 0.1,
  minpuc = 0,
  include = "",
  exclude = "",
  method = "pearson",
  maxcor = 0.71
)

Arguments

data

Matrix or data frame with incomplete data.

mincor

A scalar, numeric vector (of size ncol(data)) or numeric matrix (square, of size ncol(data) specifying the minimum threshold(s) against which the absolute correlation in the data is compared.

minpuc

A scalar, vector (of size ncol(data)) or matrix (square, of size ncol(data) specifying the minimum threshold(s) for the proportion of usable cases.

include

A string or a vector of strings containing one or more variable names from names(data). Variables specified are always included as a predictor.

exclude

A string or a vector of strings containing one or more variable names from names(data). Variables specified are always excluded as a predictor.

method

A string specifying the type of correlation. Use 'pearson' (default), 'kendall' or 'spearman'. Can be abbreviated.

maxcor

A scalar, numeric vector (of size ncol(data)) or numeric matrix (square, of size ncol(data) specifying the MAXIMUM threshold(s) against which the absolute correlation in the data is compared. this is used to help troubleshoot and resolve non-invertible predictor matrices

Value

a predictor matrix for use in mice