Impute Missing GFS Data

Additional details forthcoming....

The imputation automatically excludes variables that are 100

The imputation method for all variables is predictive mean matching (pmm).

Usage

run_impute_data(data, data.dir, Nimp = 20, Miter = 5, ...)

Arguments

data

data.frame with raw data with missing coded as "(missing)"

data.dir

directory where to save the imputed data

Nimp

number of imputed datasets to generated

Miter

number of iteration per imputed dataset to use during imputatiuon (uses mice default)

...

other arguments passed to mice

Value

a nested (by country) data.frame of mice objects with completed imputed datasets ready for extraction and recoding.

Details

A wrapper function to conducting imputation by country.

Examples

{
  # TO-DO
}
#> NULL