Estimate and Save Attrition Weights

The default set of predictors: "ANNUAL_WEIGHT1_W1", "MODE_RECRUIT_W1", "AGE_W1", "GENDER_W1", "EDUCATION_3_W1", "EMPLOYMENT_W1", "MARITAL_STATUS_W1", "RACE_PLURALITY_W1".

Usage

append_attrition_wgts(
  data,
  attr.pred = NULL,
  stabilized = TRUE,
  robust = FALSE,
  ...
)

Arguments

data

a pre-cleaned and mostly ready for analysis data.frame that will be used to construct attrition weights.

attr.pred

(character vector) default to NULL, a vector of variables to be used as predictors of staying in the study at wave 2

stabilized

(logical) default to TRUE, where the estimated attrition weight is the divisor and the numerator is the baseline rate of attrition (if FALSE, then the numerator is fixed to 1)

robust

(logical) default to FALSE, of whether to use a "robust" variant of probit regression (uses the robustbase package instead of the survey package to get weights).

...

other arguments passed to svyglm or glmrob functions

Value

a dataset with attrition weights appended

Details

Estimate (save fitted model) and append attrition weights to data.frame

Examples

{
  # TO-DO
}
#> NULL