GFS Focal Predictor Analysis

The `pc.rule` argument determines how should the number of PCs is decided: exclude PCs from analysis (default: "omit"); a fixed number ("constant"); minimunm total proportion of variance explained ("mintotal"); or minimum proportion explain by individual PCs ("mincomp"). For "constant", `pc.cutoff` must of a number between 1 and length(contemporaneous.exposures) - 1. For "mintotal", `pc.cutoff` must be strictly between 0 and 1, e.g. (0.70). For "mincomp", `pc.cutoff` must be strictly between 0 and 1, e.g. (0.01).

When standardize is TRUE, standardize the outcome and predictor using the survey-adjusted mean and variance.

When force.linear is TRUE, forces the estimation of a linear model even for binary/Likert outcomes. I think this is sometimes called as a "linear probability model."

When robust.huberM is TRUE, for (approximately) continuous outcomes, you have the option of alternatively using a "robust m-estimator" with Huber style robustness weights in addition to the complex sampling design adjustments. It's unknown whether this makes a meaningful difference, but preliminary testing suggests small differences in point estimates but sometimes dramatic changes to standard errors for reasons that are unclear to me. Could be due to a strange interaction of robustness weights, attrition weights, and post-stratified sampling weights.

gfs_run_regression_single_outcome(
  data = NULL,
  your.pred = NULL,
  your.outcome = NULL,
  covariates = NULL,
  contemporaneous.exposures = NULL,
  standardize = TRUE,
  pc.cutoff = 7,
  pc.rule = "omit",
  force.linear = FALSE,
  robust.huberM = FALSE,
  robust.tune = 1,
  res.dir = NULL,
  ...
)

Arguments

data

long data.frame must have COUNTRY and .imp columns (default: df.imp.long)

your.pred

a character string defining focal predictor (e.g., "PHYSICAL_HLTH_W1")

your.outcome

a character string defining outcome variable (e.g., "HAPPY_W2")

covariates

a character vector defining core set of covariates (default: NULL)

contemporaneous.exposures

a character vector defining set of contemporaneous exposures that are input into PCA (default: NULL)

standardize

a logical (default: TRUE) of whether to standardized the predictor and outcome

pc.cutoff

a numeric value, can either be a fixed whole number (e.g., keep 7 PC in all countries) OR a proportion (e.g., 0.50)

pc.rule

principal components analysis use rule (default: "omit")

force.linear

a logical of whether to force a linear model (default: FALSE)

robust.huberM

a logical of whether to use a robust variant of the linear regression model (default: FALSE), see below for additional details.

robust.tune

a numeric value defining the tuning parameter for the robust.huberM option.

res.dir

a character string defining directory to save results to.

...

other arguments passed to svyglm or glmrob functions

Value

a data.frame that contains the meta-analysis input results

Details

Run regression analyses for specific focal predictor using the core GFS group analysis pathway.

Examples

{
  # TO-DO
}
#> NULL