Impute Missing GFS Data

Additional details forthcoming....

The imputation automatically excludes variables that are 100

The imputation method for most variables is predictive mean matching (pmm). You can change this by using the use.log.poly option to use logistic & multinomial logistic regression for all factor (categorical) variables. Some variables use CART (classification and regression tree) method to help avoid singularities and errors in during the imputation stage. Such as factor/nominal variables with a lot of categories.

For more information about pmm see https://stefvanbuuren.name/fimd/sec-pmm.html

Usage

run_impute_data(
  data,
  data.dir,
  Nimp = 20,
  Miter = 5,
  pred.vars = NULL,
  use.log.poly = FALSE,
  ...
)

Arguments

data

data.frame with raw data with missing coded as "(missing)", must be a flat data.frame with no nesting

data.dir

directory where to save the imputed data

Nimp

number of imputed datasets to generated (default of 20)

Miter

number of iteration per imputed dataset to use during imputatiuon (uses mice default of 5)

pred.vars

(optional) character vector of variable names to to as potential predictors in missing data model (see details for more information)

use.log.poly

(FALSE) whether to use logistic/mutlinomial logistic regression for factor variables.

...

other arguments passed to mice

Value

a nested (by country) data.frame of mice objects with completed imputed datasets ready for extraction and recoding.

Details

A wrapper function to conducting imputation by country.

Examples

{
  # TO-DO
}
#> NULL